+++ packages/components/src/internal/components/samples/SampleTypeEmptyAlert.tsx
